requirements.txt
tensorflow>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
mlflow>=2.5.0
dvc>=3.25.0
streamlit>=1.25.0
plotly>=5.15.0
opencv-python>=4.8.0
pillow>=10.0.0
python-dotenv>=1.0.0